Added the option to click on an installed extension from the navbar.
This should allow users to interact with an extension after they
install it by simply knowing the name.

Since webextensions can not change the URL of the current page,
FoxPuppet with switch to the new tab that gets opened once the
extension is clicked. If no new tab is opened the current tab
will still be selected for interaction.